refactored control to separate fil
